+++ spec/integration/resource/api600/c7000/server_hardware/update_spec.rb
@@ -1,13 +1,13 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::ServerHardware
+klass = OneviewSDK::API600::C7000::ServerHardware
 RSpec.describe klass, integration: true, type: UPDATE do
-  let(:current_client) { $client_500 }
+  let(:current_client) { $client_600 }
   let(:hostname) { $secrets['server_hardware_ip'] }
 
-  include_examples 'ServerHardwareUpdateExample', 'integration api500 context', 500
+  include_examples 'ServerHardwareUpdateExample', 'integration api600 context', 600
 
-  include_examples 'ScopeHelperMethodsExample', OneviewSDK::API500::C7000::Scope do
+  include_examples 'ScopeHelperMethodsExample', OneviewSDK::API600::C7000::Scope do
     let(:item) { described_class.get_all(current_client).first }
   end
 end